add changes to navbar
