+++ offchain-network-front-end/package.json
@@ -11,8 +11,10 @@
     "font-awesome": "^4.7.0",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
+    "react-redux": "^7.2.4",
     "react-router-dom": "^5.2.0",
     "react-scripts": "4.0.3",
+    "redux": "^4.1.0",
     "web-vitals": "^1.0.1"
   },
   "scripts": {
