+++ offchain-network-front-end/src/css/erc20.css
@@ -64,3 +64,7 @@ h3 {
 	border: none;
 	color: #fff;
 }
+
+.token-address span {
+	font-size: 12px;
+}
